refactor publish helper
